// src/lib/interfaces/artist-album.ts
export interface ExternalUrls {
  spotify: string
}

export interface Artist {
  external_urls: ExternalUrls
  href: string
  id: string
  name: string
  type: string
  uri: string
}

export interface ExternalUrls2 {
  spotify: string
}

export interface Image {
  height: number
  url: string
  width: number
}

export interface Item {
  album_group?: string
  album_type: string
  artists: Artist[]
  external_urls: ExternalUrls2
  href: string
  id: string
  images: Image[]
  name: string
  release_date: string
  release_date_precision: string
  total_tracks: number
  type: string
  uri: string
}

export interface IArtistsAlbums {
  href?: string
  items: Item[]
  limit?: number
  next?: string
  offset?: number
  previous?: string
  total?: number
}